mejorando la logica del panel, busqueda de que exista la cedula, y flash messages

+++ database/migrations/2023_12_20_204700_create_users_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->unsignedInteger('roles_id');
+            $table->unsignedBigInteger('roles_id');
             $table->rememberToken();
             $table->timestamps();
             $table->foreign('roles_id')->references('id')->on('roles')
